Combine backend implementation classes into Target
There were previously classes NamespaceManager and Unparser that would
have to be extended by each backend implementation, with plans for more
to address non-expression syntax. It's easy to see that this would later
result in it being a huge pain to write each particular backend. These
have been combined into the class Target, which is where all
implementation-specific methods will live from now on. Furthermore,
there is now the abstract class ASTNode, (to be renamed to avoid
confusing it with KORE's ASTNode) which holds a Target in each instance,
and which should be extended by every syntactic construct that is added.